pyproject.toml
[project]
name = "hydra-essentials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "rich>=14.2.0",
    "torch>=2.9.0",
]
